added 'Browser#try' method and 'Browser#open?' method for review.
added 'Browser#try' method and 'Browser#open?' method for review. Created pull request in response to a question I made on stackoverflow recently which received no suitable responses for my purposes: http://stackoverflow.com/questions/35283059/how-do-i-find-out-if-a-watir-objects-browser-is-closed-or-not-after-typing-ct/35297847#35297847.About the code, #open? currently relies on #try, but if the #open? command is desired and the other is not, the #try? method's code could be extracted into the #open? method.